Bugfixes and Settings UI Tweaks
Changes:
added functions:
 updateSettingValue(id)
 @lrs[id] #access setting by its name

added the very handy (i think)
ability to refresh settings automatically in
the user interface when they are changed.
At the moment this only happens with file_export_path,
but if this is deemed usefull, then I guess others
should be added!

Now able to view export path in the
settings window.

Bugfixes:
 children of checkboxes now hide and unhide when
 the checkbox is checked by an external function
 such as when the window is opened and the variables are
 refreshed.

If there is a better way to do this (I've only just started
learning javascript due to this exporter), please feel free to
fix my mistakes (exvion and mimhotep)!

I'm beginning to think that settings should have
their own types so that it is easy to tell how
they should be displayed, and whether they
need to be updated in the ui, rather than
storing yet more lists of them. how does this sound?