More work on LuxrenderTypes
Almost 500 lines now, and is actually close to functionality now!
I think I might need some help getting tab counting and orginising
working. I was thinking of doing this in a similar way to @lrs
as a counter that can be accessed by all the classes during
the export process.

Setting up material ui, and setting ui using the html exporter
should be a cinch. How often to export this html layout is something
that I havn't really concluded. This should also allow backwards
compatability with files without too much difficulty
because you should be able generate a different to suit
the settings available. This might not turn out to be reality,
maybe there is a good reason why no-one else is trying it!?

If it does work, then, it will be really snazy!
(even though all of this should be invisible to the user)

The syntax is fairly flexible, so as to allow you guys
to lay it out in a way that is similar to how actual
html is layed out. (I havn't had the time to experiment
with the layout when defining the properties and the
settings panel).

I've also tried to clearly mark how it works
so it shouldn't be too hard to add extra functionality.