Fixed another nasty reloading bug
Previously there was a crash whenever a file was loaded or a new file was started
without having already opened up the settings editor and created the settings
in the AttributDics.

Now it is much smarter, and remaps the setting upon open or new.
On new it also recreates initial values, but that hardly matters,
because after that the remapping, the default settings file is
loaded anyway...